Extract BodyParser trait and move JsonBodyParser to own mod
